Improve test case logging
Warnings and errors were not logged to the console,
but to the test.log file instead. This has been improved
by creating a standard test logging setup function.
