+++ montblanc/tests/test_cmp_vis.py
@@ -103,15 +103,7 @@ class TestCmpVis(unittest.TestCase):
     def setUp(self):
         """ Set up each test case """
         np.random.seed(int(time.time()) & 0xFFFFFFFF)
-
-        # Add a handler that outputs INFO level logging to file
-        fh = logging.FileHandler('test.log')
-        fh.setLevel(logging.INFO)
-        ch = logging.StreamHandler()
-        ch.setLevel(logging.WARN)
-
-        montblanc.log.setLevel(logging.INFO)
-        montblanc.log.handlers = [fh, ch]
+        montblanc.setup_test_logging()
 
     def tearDown(self):
         """ Tear down each test case """
